backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
PyMuPDF==1.23.8
python-dotenv==1.0.0
openai==1.3.5
aiofiles==23.2.1
httpx==0.25.1
